push para o grande Eduu

+++ include/libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: eddos-sa <eddos-sa@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/26 13:19:10 by eddos-sa          #+#    #+#             */
-/*   Updated: 2023/12/07 18:24:47 by eddos-sa         ###   ########.fr       */
+/*   Updated: 2024/03/01 14:02:48 by eddos-sa         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	len;
+	size_t	len;
 
 	len = 0;
 	while (s[len] != '\0')
